Move roleset secrets tests to new file
